Enable tests in JPA Spec20 FAT
