+++ dev/com.ibm.ws.jpa.tests.spec20_fat/fat/src/com/ibm/ws/jpa/tests/spec20/FATSuite.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 IBM Corporation and others.
+ * Copyright (c) 2020, 2022 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -11,47 +11,17 @@
 
 package com.ibm.ws.jpa.tests.spec20;
 
-import org.junit.ClassRule;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
-import org.testcontainers.containers.JdbcDatabaseContainer;
 
-import componenttest.containers.ExternalTestServiceDockerClientStrategy;
-import componenttest.rules.repeater.FeatureReplacementAction;
-import componenttest.rules.repeater.RepeatTests;
-import componenttest.topology.database.container.DatabaseContainerFactory;
+import com.ibm.ws.jpa.tests.spec20.tests.AbstractFATSuite;
 
 @RunWith(Suite.class)
 @SuiteClasses({
                 JPA20FATSuite.class,
-//                TestOLGH9018_EJB.class,
-//                TestOLGH9018_WEB.class,
-//                TestOLGH9339_EJB.class,
-//                TestOLGH9339_WEB.class,
-//                TestOLGH10515_EJB.class,
-//                TestOLGH10515_WEB.class,
-//                TestOLGH16686_EJB.class,
-//                TestOLGH16686_WEB.class,
                 componenttest.custom.junit.runner.AlwaysPassesTest.class
 })
-public class FATSuite {
-    public final static String[] JAXB_PERMS = { "permission java.lang.RuntimePermission \"accessClassInPackage.com.sun.xml.internal.bind.v2.runtime.reflect\";",
-                                                "permission java.lang.RuntimePermission \"accessClassInPackage.com.sun.xml.internal.bind\";" };
+public class FATSuite extends AbstractFATSuite {
 
-    @ClassRule
-    public static RepeatTests r = RepeatTests.withoutModification()
-                    .andWith(FeatureReplacementAction.EE7_FEATURES())
-                    .andWith(new RepeatWithJPA20())
-                    .andWith(FeatureReplacementAction.EE9_FEATURES())
-                    .andWith(FeatureReplacementAction.EE10_FEATURES());
-
-    //Required to ensure we calculate the correct strategy each run even when
-    //switching between local and remote docker hosts.
-    static {
-        ExternalTestServiceDockerClientStrategy.setupTestcontainers();
-    }
-
-    @ClassRule
-    public static JdbcDatabaseContainer<?> testContainer = DatabaseContainerFactory.create();
 }
