+++ dev/com.ibm.ws.jpa.tests.spec20_fat/fat/src/com/ibm/ws/jpa/tests/spec20/tests/JPA20CriteriaQuery_EJB.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 IBM Corporation and others.
+ * Copyright (c) 2019 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.jpa.tests.spec20.olgh;
+package com.ibm.ws.jpa.tests.spec20.tests;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -27,7 +27,9 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.Application;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
-import com.ibm.ws.jpa.olgh10515.web.TestOLGH10515Servlet;
+import com.ibm.ws.jpa.fvt.criteriaquery.ejb.TestCriteriaQuery_EJB_SFEx_Servlet;
+import com.ibm.ws.jpa.fvt.criteriaquery.ejb.TestCriteriaQuery_EJB_SF_Servlet;
+import com.ibm.ws.jpa.fvt.criteriaquery.ejb.TestCriteriaQuery_EJB_SL_Servlet;
 import com.ibm.ws.jpa.tests.spec20.FATSuite;
 import com.ibm.ws.jpa.tests.spec20.JPAFATServletClient;
 
@@ -44,11 +46,11 @@ import componenttest.topology.utils.PrivHelper;
 
 @RunWith(FATRunner.class)
 @Mode(TestMode.FULL)
-public class TestOLGH10515_WEB extends JPAFATServletClient {
-    private final static String CONTEXT_ROOT = "olgh10515Web";
-    private final static String RESOURCE_ROOT = "test-applications/olgh10515/";
-    private final static String appFolder = "web";
-    private final static String appName = "olgh10515Web";
+public class JPA20CriteriaQuery_EJB extends JPAFATServletClient {
+    private final static String CONTEXT_ROOT = "criteriaqueryEjb";
+    private final static String RESOURCE_ROOT = "test-applications/criteriaquery/";
+    private final static String appFolder = "ejb";
+    private final static String appName = "criteriaqueryEjb";
     private final static String appNameEar = appName + ".ear";
 
     private final static Set<String> dropSet = new HashSet<String>();
@@ -58,14 +60,16 @@ public class TestOLGH10515_WEB extends JPAFATServletClient {
     private static long timestart = 0;
 
     static {
-        dropSet.add("OLGH10515_DROP_${dbvendor}.ddl");
-        createSet.add("OLGH10515_CREATE_${dbvendor}.ddl");
-        populateSet.add("OLGH10515_POPULATE_${dbvendor}.ddl");
+        dropSet.add("JPA_CRITERIAQUERY_DROP_${dbvendor}.ddl");
+        createSet.add("JPA_CRITERIAQUERY_CREATE_${dbvendor}.ddl");
+        populateSet.add("JPA_CRITERIAQUERY_POPULATE_${dbvendor}.ddl");
     }
 
-    @Server("JPA20Server")
+    @Server("JPA20CriteriaQueryEJBServer")
     @TestServlets({
-                    @TestServlet(servlet = TestOLGH10515Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH10515Servlet")
+                    @TestServlet(servlet = TestCriteriaQuery_EJB_SL_Servlet.class, path = CONTEXT_ROOT + "/" + "TestCriteriaQuery_EJB_SL_Servlet"),
+                    @TestServlet(servlet = TestCriteriaQuery_EJB_SF_Servlet.class, path = CONTEXT_ROOT + "/" + "TestCriteriaQuery_EJB_SF_Servlet"),
+                    @TestServlet(servlet = TestCriteriaQuery_EJB_SFEx_Servlet.class, path = CONTEXT_ROOT + "/" + "TestCriteriaQuery_EJB_SFEx_Servlet")
     })
     public static LibertyServer server;
 
@@ -74,7 +78,7 @@ public class TestOLGH10515_WEB extends JPAFATServletClient {
     @BeforeClass
     public static void setUp() throws Exception {
         PrivHelper.generateCustomPolicy(server, FATSuite.JAXB_PERMS);
-        bannerStart(TestOLGH10515_WEB.class);
+        bannerStart(JPA20CriteriaQuery_EJB.class);
         timestart = System.currentTimeMillis();
 
         //Get driver name
@@ -111,15 +115,20 @@ public class TestOLGH10515_WEB extends JPAFATServletClient {
     }
 
     private static void setupTestApplication() throws Exception {
+        JavaArchive ejbApp = ShrinkWrap.create(JavaArchive.class, appName + ".jar");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.fvt.criteriaquery.ejblocal");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.fvt.criteriaquery.model");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.fvt.criteriaquery.testlogic");
+        ShrinkHelper.addDirectory(ejbApp, RESOURCE_ROOT + appFolder + "/" + appName + ".jar");
+
         WebArchive webApp = ShrinkWrap.create(WebArchive.class, appName + ".war");
-        webApp.addPackages(true, "com.ibm.ws.jpa.olgh10515.model");
-        webApp.addPackages(true, "com.ibm.ws.jpa.olgh10515.testlogic");
-        webApp.addPackages(true, "com.ibm.ws.jpa.olgh10515.web");
+        webApp.addPackages(true, "com.ibm.ws.jpa.fvt.criteriaquery.ejb");
         ShrinkHelper.addDirectory(webApp, RESOURCE_ROOT + appFolder + "/" + appName + ".war");
 
         final JavaArchive testApiJar = buildTestAPIJar();
 
         final EnterpriseArchive app = ShrinkWrap.create(EnterpriseArchive.class, appNameEar);
+        app.addAsModule(ejbApp);
         app.addAsModule(webApp);
         app.addAsLibrary(testApiJar);
         ShrinkHelper.addDirectory(app, RESOURCE_ROOT + appFolder, new org.jboss.shrinkwrap.api.Filter<ArchivePath>() {
@@ -180,7 +189,7 @@ public class TestOLGH10515_WEB extends JPAFATServletClient {
             } catch (Throwable t) {
                 t.printStackTrace();
             }
-            bannerEnd(TestOLGH10515_WEB.class, timestart);
+            bannerEnd(JPA20CriteriaQuery_EJB.class, timestart);
         }
     }
 }
