+++ dev/com.ibm.ws.jpa.tests.spec20_fat/fat/src/com/ibm/ws/jpa/tests/spec20/tests/JPA20Query_EJB.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019 IBM Corporation and others.
+ * Copyright (c) 2020 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.jpa.tests.spec20;
+package com.ibm.ws.jpa.tests.spec20.tests;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -27,9 +27,11 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.Application;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
-import com.ibm.ws.jpa.cache.web.CacheTestServlet;
-import com.ibm.ws.jpa.cache.web.FindCacheTestServlet;
-import com.ibm.ws.jpa.cache.web.RefreshCacheTestServlet;
+import com.ibm.ws.jpa.query.ejb.TestQuery_EJB_SFEx_Servlet;
+import com.ibm.ws.jpa.query.ejb.TestQuery_EJB_SF_Servlet;
+import com.ibm.ws.jpa.query.ejb.TestQuery_EJB_SL_Servlet;
+import com.ibm.ws.jpa.tests.spec20.FATSuite;
+import com.ibm.ws.jpa.tests.spec20.JPAFATServletClient;
 
 import componenttest.annotation.Server;
 import componenttest.annotation.TestServlet;
@@ -44,11 +46,11 @@ import componenttest.topology.utils.PrivHelper;
 
 @RunWith(FATRunner.class)
 @Mode(TestMode.FULL)
-public class JPA20Cache_WEB extends JPAFATServletClient {
-    private final static String CONTEXT_ROOT = "cacheWeb";
-    private final static String RESOURCE_ROOT = "test-applications/cache/";
-    private final static String appFolder = "web";
-    private final static String appName = "cacheWeb";
+public class JPA20Query_EJB extends JPAFATServletClient {
+    private final static String CONTEXT_ROOT = "queryEjb";
+    private final static String RESOURCE_ROOT = "test-applications/query/";
+    private final static String appFolder = "ejb";
+    private final static String appName = "queryEjb";
     private final static String appNameEar = appName + ".ear";
 
     private final static Set<String> dropSet = new HashSet<String>();
@@ -57,16 +59,15 @@ public class JPA20Cache_WEB extends JPAFATServletClient {
     private static long timestart = 0;
 
     static {
-        dropSet.add("JPA_CACHE_DROP_${dbvendor}.ddl");
-        createSet.add("JPA_CACHE_CREATE_${dbvendor}.ddl");
+        dropSet.add("JPA20_QUERY_DROP_${dbvendor}.ddl");
+        createSet.add("JPA20_QUERY_CREATE_${dbvendor}.ddl");
     }
 
-    @Server("JPA20CacheServer")
+    @Server("JPA20Server")
     @TestServlets({
-                    @TestServlet(servlet = CacheTestServlet.class, path = CONTEXT_ROOT + "/" + "CacheTestServlet"),
-                    @TestServlet(servlet = FindCacheTestServlet.class, path = CONTEXT_ROOT + "/" + "FindCacheTestServlet"),
-                    @TestServlet(servlet = RefreshCacheTestServlet.class, path = CONTEXT_ROOT + "/" + "RefreshCacheTestServlet"),
-
+                    @TestServlet(servlet = TestQuery_EJB_SL_Servlet.class, path = CONTEXT_ROOT + "/" + "TestQuery_EJB_SL_Servlet"),
+                    @TestServlet(servlet = TestQuery_EJB_SF_Servlet.class, path = CONTEXT_ROOT + "/" + "TestQuery_EJB_SF_Servlet"),
+                    @TestServlet(servlet = TestQuery_EJB_SFEx_Servlet.class, path = CONTEXT_ROOT + "/" + "TestQuery_EJB_SFEx_Servlet")
     })
     public static LibertyServer server;
 
@@ -75,7 +76,7 @@ public class JPA20Cache_WEB extends JPAFATServletClient {
     @BeforeClass
     public static void setUp() throws Exception {
         PrivHelper.generateCustomPolicy(server, FATSuite.JAXB_PERMS);
-        bannerStart(JPA20Cache_WEB.class);
+        bannerStart(JPA20Query_EJB.class);
         timestart = System.currentTimeMillis();
 
         int appStartTimeout = server.getAppStartTimeout();
@@ -100,6 +101,8 @@ public class JPA20Cache_WEB extends JPAFATServletClient {
 
         final Set<String> ddlSet = new HashSet<String>();
 
+        System.out.println("TestQuery_EJB Setting up database tables...");
+
         ddlSet.clear();
         for (String ddlName : dropSet) {
             ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
@@ -112,18 +115,30 @@ public class JPA20Cache_WEB extends JPAFATServletClient {
         }
         executeDDL(server, ddlSet, false);
 
+//        ddlSet.clear();
+//        for (String ddlName : populateSet) {
+//            ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
+//        }
+//        executeDDL(server, ddlSet, false);
+
         setupTestApplication();
     }
 
     private static void setupTestApplication() throws Exception {
+        JavaArchive ejbApp = ShrinkWrap.create(JavaArchive.class, appName + ".jar");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.query.ejblocal");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.query.model");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.query.testlogic");
+        ShrinkHelper.addDirectory(ejbApp, RESOURCE_ROOT + appFolder + "/" + appName + ".jar");
+
         WebArchive webApp = ShrinkWrap.create(WebArchive.class, appName + ".war");
-        webApp.addPackages(true, "com.ibm.ws.jpa.cache.model");
-        webApp.addPackages(true, "com.ibm.ws.jpa.cache.web");
+        webApp.addPackages(true, "com.ibm.ws.jpa.query.ejb");
         ShrinkHelper.addDirectory(webApp, RESOURCE_ROOT + appFolder + "/" + appName + ".war");
 
         final JavaArchive testApiJar = buildTestAPIJar();
 
         final EnterpriseArchive app = ShrinkWrap.create(EnterpriseArchive.class, appNameEar);
+        app.addAsModule(ejbApp);
         app.addAsModule(webApp);
         app.addAsLibrary(testApiJar);
         ShrinkHelper.addDirectory(app, RESOURCE_ROOT + appFolder, new org.jboss.shrinkwrap.api.Filter<ArchivePath>() {
@@ -143,6 +158,11 @@ public class JPA20Cache_WEB extends JPAFATServletClient {
         Application appRecord = new Application();
         appRecord.setLocation(appNameEar);
         appRecord.setName(appName);
+//        ConfigElementList<ClassloaderElement> cel = appRecord.getClassloaders();
+//        ClassloaderElement loader = new ClassloaderElement();
+//        loader.setApiTypeVisibility("+third-party");
+////        loader.getCommonLibraryRefs().add("HibernateLib");
+//        cel.add(loader);
 
         server.setMarkToEndOfLog();
         ServerConfiguration sc = server.getServerConfiguration();
@@ -158,9 +178,19 @@ public class JPA20Cache_WEB extends JPAFATServletClient {
     @AfterClass
     public static void tearDown() throws Exception {
         try {
+            // Clean up database
+            try {
+                final Set<String> ddlSet = new HashSet<String>();
+                for (String ddlName : dropSet) {
+                    ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
+                }
+                executeDDL(server, ddlSet, true);
+            } catch (Throwable t) {
+                t.printStackTrace();
+            }
+
             server.stopServer("CWWJP9991W", // From Eclipselink drop-and-create tables option
-                              "WTRN0074E: Exception caught from before_completion synchronization operation", // RuntimeException test, expected
-                              "DSRA0080E", "DSRA0010E" // Can happen with Oracle + OpenJPA
+                              "WTRN0074E: Exception caught from before_completion synchronization operation" // RuntimeException test, expected
             );
         } finally {
             try {
@@ -174,7 +204,7 @@ public class JPA20Cache_WEB extends JPAFATServletClient {
             } catch (Throwable t) {
                 t.printStackTrace();
             }
-            bannerEnd(JPA20Cache_WEB.class, timestart);
+            bannerEnd(JPA20Query_EJB.class, timestart);
         }
     }
 }
