+++ dev/com.ibm.ws.jpa.tests.spec20_fat/fat/src/com/ibm/ws/jpa/tests/spec20/tests/JPA20Query_WEB.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019 IBM Corporation and others.
+ * Copyright (c) 2020 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.jpa.tests.spec20.olgh;
+package com.ibm.ws.jpa.tests.spec20.tests;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -27,7 +27,7 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.Application;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
-import com.ibm.ws.jpa.olgh9018.web.TestOLGH9018Servlet;
+import com.ibm.ws.jpa.query.web.TestQueryServlet;
 import com.ibm.ws.jpa.tests.spec20.FATSuite;
 import com.ibm.ws.jpa.tests.spec20.JPAFATServletClient;
 
@@ -44,11 +44,11 @@ import componenttest.topology.utils.PrivHelper;
 
 @RunWith(FATRunner.class)
 @Mode(TestMode.FULL)
-public class TestOLGH9018_WEB extends JPAFATServletClient {
-    private final static String CONTEXT_ROOT = "olgh9018Web";
-    private final static String RESOURCE_ROOT = "test-applications/olgh9018/";
+public class JPA20Query_WEB extends JPAFATServletClient {
+    private final static String CONTEXT_ROOT = "queryWeb";
+    private final static String RESOURCE_ROOT = "test-applications/query/";
     private final static String appFolder = "web";
-    private final static String appName = "olgh9018Web";
+    private final static String appName = "queryWeb";
     private final static String appNameEar = appName + ".ear";
 
     private final static Set<String> dropSet = new HashSet<String>();
@@ -57,13 +57,13 @@ public class TestOLGH9018_WEB extends JPAFATServletClient {
     private static long timestart = 0;
 
     static {
-        dropSet.add("OLGH9018_DROP_${dbvendor}.ddl");
-        createSet.add("OLGH9018_CREATE_${dbvendor}.ddl");
+        dropSet.add("JPA20_QUERY_DROP_${dbvendor}.ddl");
+        createSet.add("JPA20_QUERY_CREATE_${dbvendor}.ddl");
     }
 
     @Server("JPA20Server")
     @TestServlets({
-                    @TestServlet(servlet = TestOLGH9018Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH9018Servlet")
+                    @TestServlet(servlet = TestQueryServlet.class, path = CONTEXT_ROOT + "/" + "TestQueryServlet")
     })
     public static LibertyServer server;
 
@@ -72,8 +72,18 @@ public class TestOLGH9018_WEB extends JPAFATServletClient {
     @BeforeClass
     public static void setUp() throws Exception {
         PrivHelper.generateCustomPolicy(server, FATSuite.JAXB_PERMS);
-        bannerStart(TestOLGH9018_WEB.class);
+        bannerStart(JPA20Query_WEB.class);
         timestart = System.currentTimeMillis();
+
+        int appStartTimeout = server.getAppStartTimeout();
+        if (appStartTimeout < (120 * 1000)) {
+            server.setAppStartTimeout(120 * 1000);
+        }
+
+        int configUpdateTimeout = server.getConfigUpdateTimeout();
+        if (configUpdateTimeout < (120 * 1000)) {
+            server.setConfigUpdateTimeout(120 * 1000);
+        }
 
         //Get driver name
         server.addEnvVar("DB_DRIVER", DatabaseContainerType.valueOf(testContainer).getDriverName());
@@ -87,6 +97,8 @@ public class TestOLGH9018_WEB extends JPAFATServletClient {
 
         final Set<String> ddlSet = new HashSet<String>();
 
+        System.out.println("TestQuery_Web Setting up database tables...");
+
         ddlSet.clear();
         for (String ddlName : dropSet) {
             ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
@@ -99,14 +111,20 @@ public class TestOLGH9018_WEB extends JPAFATServletClient {
         }
         executeDDL(server, ddlSet, false);
 
+//        ddlSet.clear();
+//        for (String ddlName : populateSet) {
+//            ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
+//        }
+//        executeDDL(server, ddlSet, false);
+
         setupTestApplication();
     }
 
     private static void setupTestApplication() throws Exception {
         WebArchive webApp = ShrinkWrap.create(WebArchive.class, appName + ".war");
-        webApp.addPackages(true, "com.ibm.ws.jpa.olgh9018.model");
-        webApp.addPackages(true, "com.ibm.ws.jpa.olgh9018.testlogic");
-        webApp.addPackages(true, "com.ibm.ws.jpa.olgh9018.web");
+        webApp.addPackages(true, "com.ibm.ws.jpa.query.model");
+        webApp.addPackages(true, "com.ibm.ws.jpa.query.testlogic");
+        webApp.addPackages(true, "com.ibm.ws.jpa.query.web");
         ShrinkHelper.addDirectory(webApp, RESOURCE_ROOT + appFolder + "/" + appName + ".war");
 
         final JavaArchive testApiJar = buildTestAPIJar();
@@ -131,6 +149,11 @@ public class TestOLGH9018_WEB extends JPAFATServletClient {
         Application appRecord = new Application();
         appRecord.setLocation(appNameEar);
         appRecord.setName(appName);
+//        ConfigElementList<ClassloaderElement> cel = appRecord.getClassloaders();
+//        ClassloaderElement loader = new ClassloaderElement();
+//        loader.setApiTypeVisibility("+third-party");
+////        loader.getCommonLibraryRefs().add("HibernateLib");
+//        cel.add(loader);
 
         server.setMarkToEndOfLog();
         ServerConfiguration sc = server.getServerConfiguration();
@@ -172,7 +195,7 @@ public class TestOLGH9018_WEB extends JPAFATServletClient {
             } catch (Throwable t) {
                 t.printStackTrace();
             }
-            bannerEnd(TestOLGH9018_WEB.class, timestart);
+            bannerEnd(JPA20Query_WEB.class, timestart);
         }
     }
 }
