+++ dev/com.ibm.ws.jpa.tests.spec20_fat/fat/src/com/ibm/ws/jpa/tests/spec20/tests/olgh/TestOLGH10515_EJB.java
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.jpa.tests.spec20.olgh;
+package com.ibm.ws.jpa.tests.spec20.tests.olgh;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -27,7 +27,9 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.Application;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
-import com.ibm.ws.jpa.olgh16686.web.TestOLGH16686Servlet;
+import com.ibm.ws.jpa.olgh10515.ejb.TestOLGH10515_EJB_SFEx_Servlet;
+import com.ibm.ws.jpa.olgh10515.ejb.TestOLGH10515_EJB_SF_Servlet;
+import com.ibm.ws.jpa.olgh10515.ejb.TestOLGH10515_EJB_SL_Servlet;
 import com.ibm.ws.jpa.tests.spec20.FATSuite;
 import com.ibm.ws.jpa.tests.spec20.JPAFATServletClient;
 
@@ -44,26 +46,30 @@ import componenttest.topology.utils.PrivHelper;
 
 @RunWith(FATRunner.class)
 @Mode(TestMode.FULL)
-public class TestOLGH16686_WEB extends JPAFATServletClient {
-    private final static String CONTEXT_ROOT = "olgh16686Web";
-    private final static String RESOURCE_ROOT = "test-applications/olgh16686/";
-    private final static String appFolder = "web";
-    private final static String appName = "olgh16686Web";
+public class TestOLGH10515_EJB extends JPAFATServletClient {
+    private final static String CONTEXT_ROOT = "olgh10515Ejb";
+    private final static String RESOURCE_ROOT = "test-applications/olgh10515/";
+    private final static String appFolder = "ejb";
+    private final static String appName = "olgh10515Ejb";
     private final static String appNameEar = appName + ".ear";
 
     private final static Set<String> dropSet = new HashSet<String>();
     private final static Set<String> createSet = new HashSet<String>();
+    private final static Set<String> populateSet = new HashSet<String>();
 
     private static long timestart = 0;
 
     static {
-        dropSet.add("OLGH16686_DROP_${dbvendor}.ddl");
-        createSet.add("OLGH16686_CREATE_${dbvendor}.ddl");
+        dropSet.add("OLGH10515_DROP_${dbvendor}.ddl");
+        createSet.add("OLGH10515_CREATE_${dbvendor}.ddl");
+        populateSet.add("OLGH10515_POPULATE_${dbvendor}.ddl");
     }
 
     @Server("JPA20Server")
     @TestServlets({
-                    @TestServlet(servlet = TestOLGH16686Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH16686Servlet")
+                    @TestServlet(servlet = TestOLGH10515_EJB_SL_Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH10515_EJB_SL_Servlet"),
+                    @TestServlet(servlet = TestOLGH10515_EJB_SF_Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH10515_EJB_SF_Servlet"),
+                    @TestServlet(servlet = TestOLGH10515_EJB_SFEx_Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH10515_EJB_SFEx_Servlet")
     })
     public static LibertyServer server;
 
@@ -72,7 +78,7 @@ public class TestOLGH16686_WEB extends JPAFATServletClient {
     @BeforeClass
     public static void setUp() throws Exception {
         PrivHelper.generateCustomPolicy(server, FATSuite.JAXB_PERMS);
-        bannerStart(TestOLGH16686_WEB.class);
+        bannerStart(TestOLGH10515_EJB.class);
         timestart = System.currentTimeMillis();
 
         //Get driver name
@@ -99,19 +105,30 @@ public class TestOLGH16686_WEB extends JPAFATServletClient {
         }
         executeDDL(server, ddlSet, false);
 
+        ddlSet.clear();
+        for (String ddlName : populateSet) {
+            ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
+        }
+        executeDDL(server, ddlSet, false);
+
         setupTestApplication();
     }
 
     private static void setupTestApplication() throws Exception {
+        JavaArchive ejbApp = ShrinkWrap.create(JavaArchive.class, appName + ".jar");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.olgh10515.ejblocal");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.olgh10515.model");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.olgh10515.testlogic");
+        ShrinkHelper.addDirectory(ejbApp, RESOURCE_ROOT + appFolder + "/" + appName + ".jar");
+
         WebArchive webApp = ShrinkWrap.create(WebArchive.class, appName + ".war");
-        webApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.model");
-        webApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.testlogic");
-        webApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.web");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh10515.ejb");
         ShrinkHelper.addDirectory(webApp, RESOURCE_ROOT + appFolder + "/" + appName + ".war");
 
         final JavaArchive testApiJar = buildTestAPIJar();
 
         final EnterpriseArchive app = ShrinkWrap.create(EnterpriseArchive.class, appNameEar);
+        app.addAsModule(ejbApp);
         app.addAsModule(webApp);
         app.addAsLibrary(testApiJar);
         ShrinkHelper.addDirectory(app, RESOURCE_ROOT + appFolder, new org.jboss.shrinkwrap.api.Filter<ArchivePath>() {
@@ -172,7 +189,7 @@ public class TestOLGH16686_WEB extends JPAFATServletClient {
             } catch (Throwable t) {
                 t.printStackTrace();
             }
-            bannerEnd(TestOLGH16686_WEB.class, timestart);
+            bannerEnd(TestOLGH10515_EJB.class, timestart);
         }
     }
 }
