+++ dev/com.ibm.ws.jpa.tests.spec20_fat/fat/src/com/ibm/ws/jpa/tests/spec20/tests/olgh/TestOLGH10515_WEB.java
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.jpa.tests.spec20.olgh;
+package com.ibm.ws.jpa.tests.spec20.tests.olgh;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -27,9 +27,7 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.Application;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
-import com.ibm.ws.jpa.olgh16686.ejb.TestOLGH16686_EJB_SFEx_Servlet;
-import com.ibm.ws.jpa.olgh16686.ejb.TestOLGH16686_EJB_SF_Servlet;
-import com.ibm.ws.jpa.olgh16686.ejb.TestOLGH16686_EJB_SL_Servlet;
+import com.ibm.ws.jpa.olgh10515.web.TestOLGH10515Servlet;
 import com.ibm.ws.jpa.tests.spec20.FATSuite;
 import com.ibm.ws.jpa.tests.spec20.JPAFATServletClient;
 
@@ -46,28 +44,28 @@ import componenttest.topology.utils.PrivHelper;
 
 @RunWith(FATRunner.class)
 @Mode(TestMode.FULL)
-public class TestOLGH16686_EJB extends JPAFATServletClient {
-    private final static String CONTEXT_ROOT = "olgh16686Ejb";
-    private final static String RESOURCE_ROOT = "test-applications/olgh16686/";
-    private final static String appFolder = "ejb";
-    private final static String appName = "olgh16686Ejb";
+public class TestOLGH10515_WEB extends JPAFATServletClient {
+    private final static String CONTEXT_ROOT = "olgh10515Web";
+    private final static String RESOURCE_ROOT = "test-applications/olgh10515/";
+    private final static String appFolder = "web";
+    private final static String appName = "olgh10515Web";
     private final static String appNameEar = appName + ".ear";
 
     private final static Set<String> dropSet = new HashSet<String>();
     private final static Set<String> createSet = new HashSet<String>();
+    private final static Set<String> populateSet = new HashSet<String>();
 
     private static long timestart = 0;
 
     static {
-        dropSet.add("OLGH16686_DROP_${dbvendor}.ddl");
-        createSet.add("OLGH16686_CREATE_${dbvendor}.ddl");
+        dropSet.add("OLGH10515_DROP_${dbvendor}.ddl");
+        createSet.add("OLGH10515_CREATE_${dbvendor}.ddl");
+        populateSet.add("OLGH10515_POPULATE_${dbvendor}.ddl");
     }
 
     @Server("JPA20Server")
     @TestServlets({
-                    @TestServlet(servlet = TestOLGH16686_EJB_SL_Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH16686_EJB_SL_Servlet"),
-                    @TestServlet(servlet = TestOLGH16686_EJB_SF_Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH16686_EJB_SF_Servlet"),
-                    @TestServlet(servlet = TestOLGH16686_EJB_SFEx_Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH16686_EJB_SFEx_Servlet")
+                    @TestServlet(servlet = TestOLGH10515Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH10515Servlet")
     })
     public static LibertyServer server;
 
@@ -76,7 +74,7 @@ public class TestOLGH16686_EJB extends JPAFATServletClient {
     @BeforeClass
     public static void setUp() throws Exception {
         PrivHelper.generateCustomPolicy(server, FATSuite.JAXB_PERMS);
-        bannerStart(TestOLGH16686_EJB.class);
+        bannerStart(TestOLGH10515_WEB.class);
         timestart = System.currentTimeMillis();
 
         //Get driver name
@@ -103,24 +101,25 @@ public class TestOLGH16686_EJB extends JPAFATServletClient {
         }
         executeDDL(server, ddlSet, false);
 
+        ddlSet.clear();
+        for (String ddlName : populateSet) {
+            ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
+        }
+        executeDDL(server, ddlSet, false);
+
         setupTestApplication();
     }
 
     private static void setupTestApplication() throws Exception {
-        JavaArchive ejbApp = ShrinkWrap.create(JavaArchive.class, appName + ".jar");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.ejblocal");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.model");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.testlogic");
-        ShrinkHelper.addDirectory(ejbApp, RESOURCE_ROOT + appFolder + "/" + appName + ".jar");
-
         WebArchive webApp = ShrinkWrap.create(WebArchive.class, appName + ".war");
-        webApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.ejb");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh10515.model");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh10515.testlogic");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh10515.web");
         ShrinkHelper.addDirectory(webApp, RESOURCE_ROOT + appFolder + "/" + appName + ".war");
 
         final JavaArchive testApiJar = buildTestAPIJar();
 
         final EnterpriseArchive app = ShrinkWrap.create(EnterpriseArchive.class, appNameEar);
-        app.addAsModule(ejbApp);
         app.addAsModule(webApp);
         app.addAsLibrary(testApiJar);
         ShrinkHelper.addDirectory(app, RESOURCE_ROOT + appFolder, new org.jboss.shrinkwrap.api.Filter<ArchivePath>() {
@@ -181,7 +180,7 @@ public class TestOLGH16686_EJB extends JPAFATServletClient {
             } catch (Throwable t) {
                 t.printStackTrace();
             }
-            bannerEnd(TestOLGH16686_EJB.class, timestart);
+            bannerEnd(TestOLGH10515_WEB.class, timestart);
         }
     }
 }
