+++ dev/com.ibm.ws.jpa.tests.spec20_fat/fat/src/com/ibm/ws/jpa/tests/spec20/tests/olgh/TestOLGH16686_EJB.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2019 IBM Corporation and others.
+ * Copyright (c) 2020 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.jpa.tests.spec20;
+package com.ibm.ws.jpa.tests.spec20.tests.olgh;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -27,9 +27,11 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.Application;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
-import com.ibm.ws.jpa.fvt.criteriaquery.ejb.TestCriteriaQuery_EJB_SFEx_Servlet;
-import com.ibm.ws.jpa.fvt.criteriaquery.ejb.TestCriteriaQuery_EJB_SF_Servlet;
-import com.ibm.ws.jpa.fvt.criteriaquery.ejb.TestCriteriaQuery_EJB_SL_Servlet;
+import com.ibm.ws.jpa.olgh16686.ejb.TestOLGH16686_EJB_SFEx_Servlet;
+import com.ibm.ws.jpa.olgh16686.ejb.TestOLGH16686_EJB_SF_Servlet;
+import com.ibm.ws.jpa.olgh16686.ejb.TestOLGH16686_EJB_SL_Servlet;
+import com.ibm.ws.jpa.tests.spec20.FATSuite;
+import com.ibm.ws.jpa.tests.spec20.JPAFATServletClient;
 
 import componenttest.annotation.Server;
 import componenttest.annotation.TestServlet;
@@ -44,30 +46,28 @@ import componenttest.topology.utils.PrivHelper;
 
 @RunWith(FATRunner.class)
 @Mode(TestMode.FULL)
-public class JPA20CriteriaQuery_EJB extends JPAFATServletClient {
-    private final static String CONTEXT_ROOT = "criteriaqueryEjb";
-    private final static String RESOURCE_ROOT = "test-applications/criteriaquery/";
+public class TestOLGH16686_EJB extends JPAFATServletClient {
+    private final static String CONTEXT_ROOT = "olgh16686Ejb";
+    private final static String RESOURCE_ROOT = "test-applications/olgh16686/";
     private final static String appFolder = "ejb";
-    private final static String appName = "criteriaqueryEjb";
+    private final static String appName = "olgh16686Ejb";
     private final static String appNameEar = appName + ".ear";
 
     private final static Set<String> dropSet = new HashSet<String>();
     private final static Set<String> createSet = new HashSet<String>();
-    private final static Set<String> populateSet = new HashSet<String>();
 
     private static long timestart = 0;
 
     static {
-        dropSet.add("JPA_CRITERIAQUERY_DROP_${dbvendor}.ddl");
-        createSet.add("JPA_CRITERIAQUERY_CREATE_${dbvendor}.ddl");
-        populateSet.add("JPA_CRITERIAQUERY_POPULATE_${dbvendor}.ddl");
+        dropSet.add("OLGH16686_DROP_${dbvendor}.ddl");
+        createSet.add("OLGH16686_CREATE_${dbvendor}.ddl");
     }
 
-    @Server("JPA20CriteriaQueryEJBServer")
+    @Server("JPA20Server")
     @TestServlets({
-                    @TestServlet(servlet = TestCriteriaQuery_EJB_SL_Servlet.class, path = CONTEXT_ROOT + "/" + "TestCriteriaQuery_EJB_SL_Servlet"),
-                    @TestServlet(servlet = TestCriteriaQuery_EJB_SF_Servlet.class, path = CONTEXT_ROOT + "/" + "TestCriteriaQuery_EJB_SF_Servlet"),
-                    @TestServlet(servlet = TestCriteriaQuery_EJB_SFEx_Servlet.class, path = CONTEXT_ROOT + "/" + "TestCriteriaQuery_EJB_SFEx_Servlet")
+                    @TestServlet(servlet = TestOLGH16686_EJB_SL_Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH16686_EJB_SL_Servlet"),
+                    @TestServlet(servlet = TestOLGH16686_EJB_SF_Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH16686_EJB_SF_Servlet"),
+                    @TestServlet(servlet = TestOLGH16686_EJB_SFEx_Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH16686_EJB_SFEx_Servlet")
     })
     public static LibertyServer server;
 
@@ -76,7 +76,7 @@ public class JPA20CriteriaQuery_EJB extends JPAFATServletClient {
     @BeforeClass
     public static void setUp() throws Exception {
         PrivHelper.generateCustomPolicy(server, FATSuite.JAXB_PERMS);
-        bannerStart(JPA20CriteriaQuery_EJB.class);
+        bannerStart(TestOLGH16686_EJB.class);
         timestart = System.currentTimeMillis();
 
         //Get driver name
@@ -103,24 +103,18 @@ public class JPA20CriteriaQuery_EJB extends JPAFATServletClient {
         }
         executeDDL(server, ddlSet, false);
 
-        ddlSet.clear();
-        for (String ddlName : populateSet) {
-            ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
-        }
-        executeDDL(server, ddlSet, false);
-
         setupTestApplication();
     }
 
     private static void setupTestApplication() throws Exception {
         JavaArchive ejbApp = ShrinkWrap.create(JavaArchive.class, appName + ".jar");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.fvt.criteriaquery.ejblocal");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.fvt.criteriaquery.model");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.fvt.criteriaquery.testlogic");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.ejblocal");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.model");
+        ejbApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.testlogic");
         ShrinkHelper.addDirectory(ejbApp, RESOURCE_ROOT + appFolder + "/" + appName + ".jar");
 
         WebArchive webApp = ShrinkWrap.create(WebArchive.class, appName + ".war");
-        webApp.addPackages(true, "com.ibm.ws.jpa.fvt.criteriaquery.ejb");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.ejb");
         ShrinkHelper.addDirectory(webApp, RESOURCE_ROOT + appFolder + "/" + appName + ".war");
 
         final JavaArchive testApiJar = buildTestAPIJar();
@@ -187,7 +181,7 @@ public class JPA20CriteriaQuery_EJB extends JPAFATServletClient {
             } catch (Throwable t) {
                 t.printStackTrace();
             }
-            bannerEnd(JPA20CriteriaQuery_EJB.class, timestart);
+            bannerEnd(TestOLGH16686_EJB.class, timestart);
         }
     }
 }
