+++ dev/com.ibm.ws.jpa.tests.spec20_fat/fat/src/com/ibm/ws/jpa/tests/spec20/tests/olgh/TestOLGH16686_WEB.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 IBM Corporation and others.
+ * Copyright (c) 2020 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.jpa.tests.spec20;
+package com.ibm.ws.jpa.tests.spec20.tests.olgh;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -27,8 +27,9 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.Application;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
-import com.ibm.ws.jpa.fvt.util.tests.ejb.UtilEJBSFTestServlet;
-import com.ibm.ws.jpa.fvt.util.tests.ejb.UtilEJBSLTestServlet;
+import com.ibm.ws.jpa.olgh16686.web.TestOLGH16686Servlet;
+import com.ibm.ws.jpa.tests.spec20.FATSuite;
+import com.ibm.ws.jpa.tests.spec20.JPAFATServletClient;
 
 import componenttest.annotation.Server;
 import componenttest.annotation.TestServlet;
@@ -43,11 +44,11 @@ import componenttest.topology.utils.PrivHelper;
 
 @RunWith(FATRunner.class)
 @Mode(TestMode.FULL)
-public class JPA20Util_EJB extends JPAFATServletClient {
-    private final static String CONTEXT_ROOT = "UtilEJB";
-    private final static String RESOURCE_ROOT = "test-applications/util/";
-    private final static String appFolder = "UtilEJB.ear";
-    private final static String appName = "UtilEJB";
+public class TestOLGH16686_WEB extends JPAFATServletClient {
+    private final static String CONTEXT_ROOT = "olgh16686Web";
+    private final static String RESOURCE_ROOT = "test-applications/olgh16686/";
+    private final static String appFolder = "web";
+    private final static String appName = "olgh16686Web";
     private final static String appNameEar = appName + ".ear";
 
     private final static Set<String> dropSet = new HashSet<String>();
@@ -56,15 +57,13 @@ public class JPA20Util_EJB extends JPAFATServletClient {
     private static long timestart = 0;
 
     static {
-        dropSet.add("JPA20_UTIL_DROP_${dbvendor}.ddl");
-        createSet.add("JPA20_UTIL_CREATE_${dbvendor}.ddl");
+        dropSet.add("OLGH16686_DROP_${dbvendor}.ddl");
+        createSet.add("OLGH16686_CREATE_${dbvendor}.ddl");
     }
 
-    @Server("JPA20UtilEJBServer")
+    @Server("JPA20Server")
     @TestServlets({
-                    @TestServlet(servlet = UtilEJBSLTestServlet.class, path = CONTEXT_ROOT + "/" + "UtilEJBSLTestServlet"),
-                    @TestServlet(servlet = UtilEJBSFTestServlet.class, path = CONTEXT_ROOT + "/" + "UtilEJBSFTestServlet"),
-
+                    @TestServlet(servlet = TestOLGH16686Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH16686Servlet")
     })
     public static LibertyServer server;
 
@@ -73,18 +72,8 @@ public class JPA20Util_EJB extends JPAFATServletClient {
     @BeforeClass
     public static void setUp() throws Exception {
         PrivHelper.generateCustomPolicy(server, FATSuite.JAXB_PERMS);
-        bannerStart(JPA20Util_EJB.class);
+        bannerStart(TestOLGH16686_WEB.class);
         timestart = System.currentTimeMillis();
-
-        int appStartTimeout = server.getAppStartTimeout();
-        if (appStartTimeout < (120 * 1000)) {
-            server.setAppStartTimeout(120 * 1000);
-        }
-
-        int configUpdateTimeout = server.getConfigUpdateTimeout();
-        if (configUpdateTimeout < (120 * 1000)) {
-            server.setConfigUpdateTimeout(120 * 1000);
-        }
 
         //Get driver name
         server.addEnvVar("DB_DRIVER", DatabaseContainerType.valueOf(testContainer).getDriverName());
@@ -98,8 +87,6 @@ public class JPA20Util_EJB extends JPAFATServletClient {
 
         final Set<String> ddlSet = new HashSet<String>();
 
-        System.out.println("Setting up database tables...");
-
         ddlSet.clear();
         for (String ddlName : dropSet) {
             ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
@@ -112,30 +99,19 @@ public class JPA20Util_EJB extends JPAFATServletClient {
         }
         executeDDL(server, ddlSet, false);
 
-//        ddlSet.clear();
-//        for (String ddlName : populateSet) {
-//            ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
-//        }
-//        executeDDL(server, ddlSet, false);
-
         setupTestApplication();
     }
 
     private static void setupTestApplication() throws Exception {
-        JavaArchive ejbApp = ShrinkWrap.create(JavaArchive.class, appName + ".jar");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.fvt.util.ejblocal");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.fvt.util.entities");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.fvt.util.testlogic");
-        ShrinkHelper.addDirectory(ejbApp, RESOURCE_ROOT + appFolder + "/" + appName + ".jar");
-
         WebArchive webApp = ShrinkWrap.create(WebArchive.class, appName + ".war");
-        webApp.addPackages(true, "com.ibm.ws.jpa.fvt.util.tests.ejb");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.model");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.testlogic");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh16686.web");
         ShrinkHelper.addDirectory(webApp, RESOURCE_ROOT + appFolder + "/" + appName + ".war");
 
         final JavaArchive testApiJar = buildTestAPIJar();
 
         final EnterpriseArchive app = ShrinkWrap.create(EnterpriseArchive.class, appNameEar);
-        app.addAsModule(ejbApp);
         app.addAsModule(webApp);
         app.addAsLibrary(testApiJar);
         ShrinkHelper.addDirectory(app, RESOURCE_ROOT + appFolder, new org.jboss.shrinkwrap.api.Filter<ArchivePath>() {
@@ -196,7 +172,7 @@ public class JPA20Util_EJB extends JPAFATServletClient {
             } catch (Throwable t) {
                 t.printStackTrace();
             }
-            bannerEnd(JPA20Util_EJB.class, timestart);
+            bannerEnd(TestOLGH16686_WEB.class, timestart);
         }
     }
 }
