+++ dev/com.ibm.ws.jpa.tests.spec20_fat/fat/src/com/ibm/ws/jpa/tests/spec20/tests/olgh/TestOLGH9018_WEB.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 IBM Corporation and others.
+ * Copyright (c) 2019 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.jpa.tests.spec20;
+package com.ibm.ws.jpa.tests.spec20.tests.olgh;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -27,7 +27,9 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.Application;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
-import com.ibm.ws.jpa.fvt.derivedidentity.tests.web.DerivedIdentityWebTestServlet;
+import com.ibm.ws.jpa.olgh9018.web.TestOLGH9018Servlet;
+import com.ibm.ws.jpa.tests.spec20.FATSuite;
+import com.ibm.ws.jpa.tests.spec20.JPAFATServletClient;
 
 import componenttest.annotation.Server;
 import componenttest.annotation.TestServlet;
@@ -42,11 +44,11 @@ import componenttest.topology.utils.PrivHelper;
 
 @RunWith(FATRunner.class)
 @Mode(TestMode.FULL)
-public class JPA20DerivedIdentity_WEB extends JPAFATServletClient {
-    private final static String CONTEXT_ROOT = "DerivedIdentityWeb";
-    private final static String RESOURCE_ROOT = "test-applications/derivedIdentity/";
-    private final static String appFolder = "apps/DerivedIdentityWeb.ear";
-    private final static String appName = "DerivedIdentityWeb";
+public class TestOLGH9018_WEB extends JPAFATServletClient {
+    private final static String CONTEXT_ROOT = "olgh9018Web";
+    private final static String RESOURCE_ROOT = "test-applications/olgh9018/";
+    private final static String appFolder = "web";
+    private final static String appName = "olgh9018Web";
     private final static String appNameEar = appName + ".ear";
 
     private final static Set<String> dropSet = new HashSet<String>();
@@ -55,13 +57,13 @@ public class JPA20DerivedIdentity_WEB extends JPAFATServletClient {
     private static long timestart = 0;
 
     static {
-        dropSet.add("JPA20_DERIVEDIDENTITY_DROP_${dbvendor}.ddl");
-        createSet.add("JPA20_DERIVEDIDENTITY_CREATE_${dbvendor}.ddl");
+        dropSet.add("OLGH9018_DROP_${dbvendor}.ddl");
+        createSet.add("OLGH9018_CREATE_${dbvendor}.ddl");
     }
 
-    @Server("JPA20DerivedIdentityWebServer")
+    @Server("JPA20Server")
     @TestServlets({
-                    @TestServlet(servlet = DerivedIdentityWebTestServlet.class, path = CONTEXT_ROOT + "/" + "DerivedIdentityWebTestServlet")
+                    @TestServlet(servlet = TestOLGH9018Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH9018Servlet")
     })
     public static LibertyServer server;
 
@@ -70,18 +72,8 @@ public class JPA20DerivedIdentity_WEB extends JPAFATServletClient {
     @BeforeClass
     public static void setUp() throws Exception {
         PrivHelper.generateCustomPolicy(server, FATSuite.JAXB_PERMS);
-        bannerStart(JPA20DerivedIdentity_WEB.class);
+        bannerStart(TestOLGH9018_WEB.class);
         timestart = System.currentTimeMillis();
-
-        int appStartTimeout = server.getAppStartTimeout();
-        if (appStartTimeout < (120 * 1000)) {
-            server.setAppStartTimeout(120 * 1000);
-        }
-
-        int configUpdateTimeout = server.getConfigUpdateTimeout();
-        if (configUpdateTimeout < (120 * 1000)) {
-            server.setConfigUpdateTimeout(120 * 1000);
-        }
 
         //Get driver name
         server.addEnvVar("DB_DRIVER", DatabaseContainerType.valueOf(testContainer).getDriverName());
@@ -95,8 +87,6 @@ public class JPA20DerivedIdentity_WEB extends JPAFATServletClient {
 
         final Set<String> ddlSet = new HashSet<String>();
 
-        System.out.println("Setting up database tables...");
-
         ddlSet.clear();
         for (String ddlName : dropSet) {
             ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
@@ -109,20 +99,14 @@ public class JPA20DerivedIdentity_WEB extends JPAFATServletClient {
         }
         executeDDL(server, ddlSet, false);
 
-//        ddlSet.clear();
-//        for (String ddlName : populateSet) {
-//            ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
-//        }
-//        executeDDL(server, ddlSet, false);
-
         setupTestApplication();
     }
 
     private static void setupTestApplication() throws Exception {
         WebArchive webApp = ShrinkWrap.create(WebArchive.class, appName + ".war");
-        webApp.addPackages(true, "com.ibm.ws.jpa.commonentities.datamodel");
-        webApp.addPackages(true, "com.ibm.ws.jpa.fvt.derivedidentity.testlogic");
-        webApp.addPackages(true, "com.ibm.ws.jpa.fvt.derivedidentity.tests.web");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh9018.model");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh9018.testlogic");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh9018.web");
         ShrinkHelper.addDirectory(webApp, RESOURCE_ROOT + appFolder + "/" + appName + ".war");
 
         final JavaArchive testApiJar = buildTestAPIJar();
@@ -188,7 +172,7 @@ public class JPA20DerivedIdentity_WEB extends JPAFATServletClient {
             } catch (Throwable t) {
                 t.printStackTrace();
             }
-            bannerEnd(JPA20DerivedIdentity_WEB.class, timestart);
+            bannerEnd(TestOLGH9018_WEB.class, timestart);
         }
     }
 }
