+++ dev/com.ibm.ws.jpa.tests.spec20_fat/fat/src/com/ibm/ws/jpa/tests/spec20/tests/olgh/TestOLGH9339_WEB.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021 IBM Corporation and others.
+ * Copyright (c) 2019 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -9,7 +9,7 @@
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 
-package com.ibm.ws.jpa.tests.spec20;
+package com.ibm.ws.jpa.tests.spec20.tests.olgh;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -27,9 +27,9 @@ import org.testcontainers.containers.JdbcDatabaseContainer;
 import com.ibm.websphere.simplicity.ShrinkHelper;
 import com.ibm.websphere.simplicity.config.Application;
 import com.ibm.websphere.simplicity.config.ServerConfiguration;
-import com.ibm.ws.jpa.fvt.derivedidentity.tests.ejb.DerivedIdentityEJBSFEXTestServlet;
-import com.ibm.ws.jpa.fvt.derivedidentity.tests.ejb.DerivedIdentityEJBSFTestServlet;
-import com.ibm.ws.jpa.fvt.derivedidentity.tests.ejb.DerivedIdentityEJBSLTestServlet;
+import com.ibm.ws.jpa.olgh9339.web.TestOLGH9339Servlet;
+import com.ibm.ws.jpa.tests.spec20.FATSuite;
+import com.ibm.ws.jpa.tests.spec20.JPAFATServletClient;
 
 import componenttest.annotation.Server;
 import componenttest.annotation.TestServlet;
@@ -44,28 +44,28 @@ import componenttest.topology.utils.PrivHelper;
 
 @RunWith(FATRunner.class)
 @Mode(TestMode.FULL)
-public class JPA20DerivedIdentity_EJB extends JPAFATServletClient {
-    private final static String CONTEXT_ROOT = "DerivedIdentityEJB";
-    private final static String RESOURCE_ROOT = "test-applications/derivedIdentity/";
-    private final static String appFolder = "apps/DerivedIdentityEJB.ear";
-    private final static String appName = "DerivedIdentityEJB";
+public class TestOLGH9339_WEB extends JPAFATServletClient {
+    private final static String CONTEXT_ROOT = "olgh9339Web";
+    private final static String RESOURCE_ROOT = "test-applications/olgh9339/";
+    private final static String appFolder = "web";
+    private final static String appName = "olgh9339Web";
     private final static String appNameEar = appName + ".ear";
 
     private final static Set<String> dropSet = new HashSet<String>();
     private final static Set<String> createSet = new HashSet<String>();
+    private final static Set<String> populateSet = new HashSet<String>();
 
     private static long timestart = 0;
 
     static {
-        dropSet.add("JPA20_DERIVEDIDENTITY_DROP_${dbvendor}.ddl");
-        createSet.add("JPA20_DERIVEDIDENTITY_CREATE_${dbvendor}.ddl");
+        dropSet.add("OLGH9339_DROP_${dbvendor}.ddl");
+        createSet.add("OLGH9339_CREATE_${dbvendor}.ddl");
+        populateSet.add("OLGH9339_POPULATE_${dbvendor}.ddl");
     }
 
-    @Server("JPA20DerivedIdentityEJBServer")
+    @Server("JPA20Server")
     @TestServlets({
-                    @TestServlet(servlet = DerivedIdentityEJBSLTestServlet.class, path = CONTEXT_ROOT + "/" + "DerivedIdentityEJBSLTestServlet"),
-                    @TestServlet(servlet = DerivedIdentityEJBSFTestServlet.class, path = CONTEXT_ROOT + "/" + "DerivedIdentityEJBSFTestServlet"),
-                    @TestServlet(servlet = DerivedIdentityEJBSFEXTestServlet.class, path = CONTEXT_ROOT + "/" + "DerivedIdentityEJBSFEXTestServlet")
+                    @TestServlet(servlet = TestOLGH9339Servlet.class, path = CONTEXT_ROOT + "/" + "TestOLGH9339Servlet")
     })
     public static LibertyServer server;
 
@@ -74,18 +74,8 @@ public class JPA20DerivedIdentity_EJB extends JPAFATServletClient {
     @BeforeClass
     public static void setUp() throws Exception {
         PrivHelper.generateCustomPolicy(server, FATSuite.JAXB_PERMS);
-        bannerStart(JPA20DerivedIdentity_EJB.class);
+        bannerStart(TestOLGH9339_WEB.class);
         timestart = System.currentTimeMillis();
-
-        int appStartTimeout = server.getAppStartTimeout();
-        if (appStartTimeout < (120 * 1000)) {
-            server.setAppStartTimeout(120 * 1000);
-        }
-
-        int configUpdateTimeout = server.getConfigUpdateTimeout();
-        if (configUpdateTimeout < (120 * 1000)) {
-            server.setConfigUpdateTimeout(120 * 1000);
-        }
 
         //Get driver name
         server.addEnvVar("DB_DRIVER", DatabaseContainerType.valueOf(testContainer).getDriverName());
@@ -99,8 +89,6 @@ public class JPA20DerivedIdentity_EJB extends JPAFATServletClient {
 
         final Set<String> ddlSet = new HashSet<String>();
 
-        System.out.println("Setting up database tables...");
-
         ddlSet.clear();
         for (String ddlName : dropSet) {
             ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
@@ -113,30 +101,25 @@ public class JPA20DerivedIdentity_EJB extends JPAFATServletClient {
         }
         executeDDL(server, ddlSet, false);
 
-//        ddlSet.clear();
-//        for (String ddlName : populateSet) {
-//            ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
-//        }
-//        executeDDL(server, ddlSet, false);
+        ddlSet.clear();
+        for (String ddlName : populateSet) {
+            ddlSet.add(ddlName.replace("${dbvendor}", getDbVendor().name()));
+        }
+        executeDDL(server, ddlSet, false);
 
         setupTestApplication();
     }
 
     private static void setupTestApplication() throws Exception {
-        final JavaArchive ejbApp = ShrinkWrap.create(JavaArchive.class, appName + ".jar");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.commonentities.datamodel");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.fvt.derivedidentity.ejblocal");
-        ejbApp.addPackages(true, "com.ibm.ws.jpa.fvt.derivedidentity.testlogic");
-        ShrinkHelper.addDirectory(ejbApp, RESOURCE_ROOT + appFolder + "/" + appName + ".jar");
-
         WebArchive webApp = ShrinkWrap.create(WebArchive.class, appName + ".war");
-        webApp.addPackages(true, "com.ibm.ws.jpa.fvt.derivedidentity.tests.ejb");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh9339.model");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh9339.testlogic");
+        webApp.addPackages(true, "com.ibm.ws.jpa.olgh9339.web");
         ShrinkHelper.addDirectory(webApp, RESOURCE_ROOT + appFolder + "/" + appName + ".war");
 
         final JavaArchive testApiJar = buildTestAPIJar();
 
         final EnterpriseArchive app = ShrinkWrap.create(EnterpriseArchive.class, appNameEar);
-        app.addAsModule(ejbApp);
         app.addAsModule(webApp);
         app.addAsLibrary(testApiJar);
         ShrinkHelper.addDirectory(app, RESOURCE_ROOT + appFolder, new org.jboss.shrinkwrap.api.Filter<ArchivePath>() {
@@ -197,7 +180,7 @@ public class JPA20DerivedIdentity_EJB extends JPAFATServletClient {
             } catch (Throwable t) {
                 t.printStackTrace();
             }
-            bannerEnd(JPA20DerivedIdentity_EJB.class, timestart);
+            bannerEnd(TestOLGH9339_WEB.class, timestart);
         }
     }
 }
